pyproject.toml
[tool.poetry]
name = "ubivision-cluster-server"
version = "0.1.0"
description = ""
authors = ["TeaBear5 <tyler.bair@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
